chore(context): rename context subfolders to lowercase
