+++ src/context/index.js
@@ -1,7 +1,7 @@
-export * from './_centra'
-export * from './_cms'
-export * from './_global'
-export * from './_i18n'
+export * from './centra'
+export * from './cms'
+export * from './global'
+export * from './i18n'
 
 /** @todo Replace mocked api calls with real implementation. */
 export { fetchProducts } from './storybook.index'
