+++ src/context/storybook.index.js
@@ -1,10 +1,10 @@
 import { sleep } from '@noaignite/utils'
 import { filter, products } from '~/api/__mock__'
 
-export * from './_centra'
-export * from './_cms'
-export * from './_global'
-export * from './_i18n'
+export * from './centra'
+export * from './cms'
+export * from './global'
+export * from './i18n'
 
 export async function fetchProducts(query) {
   await sleep(500)
